feat: create vote hooks module
